refactor: FirebaseMessaging 의존성 MessageHandler로 분리 #37
