+++ meongcare/src/main/java/com/meongcare/infra/message/MessageHandler.java
@@ -1,0 +1,7 @@
+package com.meongcare.infra.message;
+
+public interface MessageHandler {
+
+    void sendMessage(String title, String body, String logoImageUrl, String fcmToken);
+
+}
